refactor: Payment 검증 후 PG사 별 결제 인증 및 Kafka 이벤트 발행 시퀀스 PaymentService에 재작성

+++ payment-service/src/main/java/fintech/domain/service/PaymentProcessor.java
@@ -1,7 +1,9 @@
 package fintech.domain.service;
 
+import java.math.BigDecimal;
+
 public interface PaymentProcessor {
 
     // 실 결제 승인 요청
-    String pay(Long amount, String orderId);
+    String pay(BigDecimal amount, String orderId);
 }
